Create interface nodes from operation nodes
Operation nodes are now responsible for creating their own interface
nodes in the graph. Nodes now need to store a pointer to the graph and
so that creating interface nodes becomes possible.

Also added a few methods exposing input and output interfaces from
nodes, generalizing the concept of interface to other types of nodes.